+++ hr-api/src/main/java/com/digdes/controllers/DepartmentTypeApiController.java
@@ -2,8 +2,10 @@ package com.digdes.controllers;
 
 import com.digdes.dto.DepartmentTypeDto;
 ;
+import com.digdes.exceptions.EntityDeleteException;
 import com.digdes.services.DepartmentTypeDataService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,34 +21,50 @@ public class DepartmentTypeApiController{
     private DepartmentTypeDataService dataService;
 
     @PostMapping(path = "/create")
-    public DepartmentTypeDto create(@RequestBody DepartmentTypeDto info) {
-        return dataService.save(info);
+    public ResponseEntity<?> create(@RequestBody DepartmentTypeDto info) {
+        if (info.getId() != null)
+            return ResponseEntity.badRequest().body("Department type id must be null for create. Use update method");
+        return ResponseEntity.ok(dataService.create(info));
     }
 
     @PutMapping(path = "/update")
-    public DepartmentTypeDto update(@RequestBody DepartmentTypeDto info) {
-        return dataService.save(info);
+    public ResponseEntity<?> update(@RequestBody DepartmentTypeDto info) {
+        if (info.getId() == null)
+            return ResponseEntity.badRequest().body("Department type id must not be null for update. Use create method");
+        if (info.getName() == null)
+            return ResponseEntity.badRequest().body("Department type name cannot be null");
+        return ResponseEntity.ok(dataService.update(info));
     }
 
     @DeleteMapping(path = "/delete")
-    public boolean delete(@RequestBody DepartmentTypeDto info) {
-        return dataService.delete(info);
+    public ResponseEntity<?> delete(@RequestBody DepartmentTypeDto info) {
+        if (info.getId() == null)
+            return ResponseEntity.badRequest().body("Department type id must not be null for delete");
+        try {
+            Boolean result = dataService.delete(info);
+            return ResponseEntity.ok(result);
+        }
+        catch (EntityDeleteException e){
+            return ResponseEntity.internalServerError().body(e.getMessage());
+        }
     }
 
     @PostMapping(path = "/find")
-    public List<DepartmentTypeDto> find(@RequestBody DepartmentTypeDto searchRequest) {
-        return dataService.find(searchRequest);
+    public ResponseEntity<?> find(@RequestBody DepartmentTypeDto searchRequest) {
+        if (searchRequest.getId() == null && searchRequest.getName() == null)
+            return ResponseEntity.badRequest().body("Id and name fields cannot be null");
+        return ResponseEntity.ok(dataService.find(searchRequest));
     }
 
     @GetMapping(path = "/get/{id}")
-    public DepartmentTypeDto get(@PathVariable Long id) {
+    public ResponseEntity<?> get(@PathVariable Long id) {
         Optional<DepartmentTypeDto> type = dataService.get(id);
-        return type.orElseGet(DepartmentTypeDto::new);
+        return ResponseEntity.ok(type.orElseGet(DepartmentTypeDto::new));
     }
 
     @GetMapping(path = "/all")
-    public List<DepartmentTypeDto> getAll() {
-        return dataService.getAll();
+    public ResponseEntity<?> getAll() {
+        return ResponseEntity.ok(dataService.getAll());
     }
 
     public boolean loadFromFile(String fileName) {
